+++ pb/pb_migrations/1746727610_updated_tickets.js
@@ -1,0 +1,11 @@
+/// <reference path="../pb_data/types.d.ts" />
+migrate((app) => {
+  const collection = app.findCollectionByNameOrId("pbc_3306545694")
+
+  // update collection data
+  unmarshal({
+    "createRule": "@request.auth.collectionName = \"customers\""
+  }, collection)
+
+  return app.save(collection)
+})