+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.0.8
+
+### Features
+
+- Customer can submit ticket after logged in
+- Add sonner toast
+
+### Fixes
+
+- Tickets API rules won't allow customer to submit ticket
+
 ## 0.0.7
 
 ### Features